backend/requirements.txt
requests>=2.31.0
flask>=2.3.0
flask-sqlalchemy>=3.0.0
flask-migrate>=4.0.0
flask-jwt-extended>=4.5.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
deepface>=0.0.75
tensorflow>=2.14.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0